Add "complete upload" logic
