+++ services/barn-ui/package.json
@@ -42,6 +42,7 @@
 		"vite": "^5.4.12"
 	},
 	"dependencies": {
+		"@aws-sdk/client-s3": "^3.735.0",
 		"@floating-ui/dom": "^1.6.12",
 		"filesize": "^10.1.6",
 		"hls.js": "^1.5.17"
