+++ services/silo-api/src/routes/upload/cloud.rs
@@ -1,17 +1,22 @@
 use std::sync::Arc;
 
-use axum::{debug_handler, extract::State, Extension, Json};
-use db::User;
+use aws_sdk_s3::types::{
+    builders::CompletedMultipartUploadBuilder, CompletedMultipartUpload, CompletedPart,
+};
+use axum::{extract::State, Extension, Json};
+use db::{User, Video};
 use reqwest::StatusCode;
 use serde::{Deserialize, Serialize};
 
 use crate::AppState;
+use queue::queue::Message::ProcessRawVideoIntoStream;
 
 #[derive(Deserialize)]
 pub struct InitUploadRequest {
     parts: i32,
     key: String,
     content_type: String,
+    title: Option<String>,
 }
 
 #[derive(Serialize)]
@@ -24,27 +29,40 @@ struct PartUrl {
 #[derive(Serialize)]
 pub struct InitUploadResponse {
     upload_id: String,
+    video_id: String,
+    key: String,
     part_urls: Vec<PartUrl>,
 }
 
-#[debug_handler]
 /// Initializes a multipart upload to R2
 pub async fn init_upload(
     State(state): State<Arc<AppState>>,
     Extension(user): Extension<Option<User>>,
     Json(request): Json<InitUploadRequest>,
 ) -> Result<Json<InitUploadResponse>, StatusCode> {
+    // Requires a logged in user
+    let user = user.ok_or(StatusCode::UNAUTHORIZED)?;
+    // Get our bucket
     let bucket = state
         .config
         .upload_bucket
         .clone()
         .ok_or(StatusCode::INTERNAL_SERVER_ERROR)?;
 
+    // TODO: Construct the key to be in a users directory
+    // Example key for file `example_video.mp4` for user id `1234`
+    //  `uploads/1234/video_id.mp4`
+    let video_id = Video::gen_id();
+    // Get the file extension from the original key
+    let extension = request.key.split('.').last().unwrap_or("");
+    let key = format!("uploads/{}/{}.{}", user.id, video_id, extension);
+
+    // Start multipart upload on R2's side
     let start_upload_output = state
         .s3_client
         .create_multipart_upload()
         .bucket(&bucket)
-        .key(&request.key)
+        .key(&key)
         .content_type(&request.content_type)
         .send()
         .await
@@ -53,11 +71,13 @@ pub async fn init_upload(
             StatusCode::INTERNAL_SERVER_ERROR
         })?;
 
+    // Parse the upload ID out of the R2 output for downstream use while uploading
     let upload_id = start_upload_output
         .upload_id()
         .ok_or(StatusCode::INTERNAL_SERVER_ERROR)?
         .to_string();
 
+    // Generate presigned links for each part
     let mut part_urls = Vec::new();
     for part_number in 1..=request.parts {
         // Construct Duration
@@ -73,7 +93,7 @@ pub async fn init_upload(
             .s3_client
             .upload_part()
             .bucket(&bucket)
-            .key(&request.key)
+            .key(&key)
             .upload_id(&upload_id)
             .part_number(part_number)
             .presigned(config)
@@ -89,10 +109,102 @@ pub async fn init_upload(
         });
     }
 
+    // Initialize the video in the database
+    let video = Video::create(
+        &state.db,
+        Some(video_id),
+        user.id,
+        request.title,
+        Some(key.clone()),
+    )
+    .await
+    .map_err(|e| {
+        tracing::error!("Could not initialize video in database {}", e);
+        StatusCode::INTERNAL_SERVER_ERROR
+    })?;
+
     Ok(Json(InitUploadResponse {
         upload_id,
         part_urls,
+        key,
+        video_id: video.id,
     }))
 }
 
+#[derive(Deserialize)]
+pub struct Parts {
+    etag: String,
+    number: i32,
+}
+
+#[derive(Deserialize)]
+pub struct CompleteUploadRequest {
+    upload_id: String,
+    video_id: String,
+    key: String,
+    completed_parts: Vec<Parts>,
+}
+
 // TODO: Write route for completing multipart upload
+/// Completes a multipart upload to R2
+pub async fn complete_upload(
+    State(state): State<Arc<AppState>>,
+    Extension(user): Extension<Option<User>>,
+    Json(request): Json<CompleteUploadRequest>,
+) -> Result<StatusCode, StatusCode> {
+    // User required
+    let _user = user.ok_or(StatusCode::UNAUTHORIZED)?;
+    // First, let R2 know we've completed the upload
+    let bucket = state
+        .config
+        .upload_bucket
+        .clone()
+        .ok_or(StatusCode::INTERNAL_SERVER_ERROR)?;
+
+    let serialized_completed_parts = request
+        .completed_parts
+        .iter()
+        .map(|part| {
+            CompletedPart::builder()
+                .e_tag(&part.etag)
+                .part_number(part.number)
+                .build()
+        })
+        .collect();
+
+    let completed_upload = CompletedMultipartUpload::builder()
+        .set_parts(Some(serialized_completed_parts))
+        .build();
+
+    let _completed_parts = state
+        .s3_client
+        .complete_multipart_upload()
+        .bucket(&bucket)
+        .key(&request.key)
+        .upload_id(&request.upload_id)
+        .multipart_upload(completed_upload)
+        .send()
+        .await
+        .map_err(|e| {
+            tracing::error!("Could not complete multipart upload {}", e);
+            StatusCode::INTERNAL_SERVER_ERROR
+        })?;
+
+    // Create and push video processing job to queue
+    state
+        .queue
+        .push(
+            ProcessRawVideoIntoStream {
+                video_id: request.video_id,
+            },
+            queue::job::PostgresJobStatus::Queued,
+            None, // Schedule for immediate processing
+        )
+        .await
+        .map_err(|e| {
+            tracing::error!("Failed to queue video processing job: {}", e);
+            StatusCode::INTERNAL_SERVER_ERROR
+        })?;
+
+    Ok(StatusCode::ACCEPTED)
+}
